addressed comments from @wbpcode
